FE-21 Refactor : ProfileInfo코드 내 모듈화 1. ProfileInfo 내 버튼부분을 모듈화 했습니다. 2. CommonWrapper를 적용했습니다.
