+++ src/pages/MyProfile.jsx
@@ -1,13 +1,13 @@
-import ProfileInfo from "../components/modules/ProfileInfo/ProfileInfo";
-import ForSale from "../components/modules/ForSale/ForSale";
-import Post from "../components/modules/Post/Post";
+import ProfileInfo from '../components/modules/ProfileInfo/ProfileInfo';
+import ForSale from '../components/modules/ForSale/ForSale';
+import ProfilePost from '../components/modules/ProfilePost/ProfilePost';
 
 const MyProfile = () => {
   return (
     <>
       <ProfileInfo />
       <ForSale />
-      <Post />
+      <ProfilePost />
     </>
   );
 };
