+++ src/pages/YourProfile.jsx
@@ -1,11 +1,13 @@
-import ProfileInfo from "../components/modules/ProfileInfo/ProfileInfo";
-import ForSale from "../components/modules/ForSale/ForSale";
+import ProfileInfo from '../components/modules/ProfileInfo/ProfileInfo';
+import ForSale from '../components/modules/ForSale/ForSale';
+import ProfilePost from '../components/modules/ProfilePost/ProfilePost';
 
 const YourProfile = () => {
   return (
     <>
       <ProfileInfo />
       <ForSale />
+      <ProfilePost />
     </>
   );
 };
